# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DiscoveryJobRequest": {
        "additionalProperties": false,
        "properties": {
          "criteria": {
            "$ref": "#/components/schemas/DiscoveryCriteria"
          },
          "kind": {
            "const": "discovery",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "criteria"
        ],
        "title": "DiscoveryJobRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "collection_id": {
            "format": "uuid",
            "title": "Collection Id",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "reused": {
            "title": "Reused",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "retry_wait",
              "completed",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "collection_id",
          "status",
          "reused"
        ],
        "title": "JobCreatedResponse",
        "type": "object"
      },
      "JobDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt_count": {
            "minimum": 0.0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/JobAttemptResponse"
            },
            "title": "Attempts",
            "type": "array"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "checkpoint": {
            "$ref": "#/components/schemas/JobCheckpointResponse"
          },
          "collection_id": {
            "format": "uuid",
            "title": "Collection Id",
            "type": "string"
          },
          "coverage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "criteria": {
            "additionalProperties": true,
            "title": "Criteria",
            "type": "object"
          },
          "cursor_invalid": {
            "title": "Cursor Invalid",
            "type": "boolean"
          },
          "environment": {
            "enum": [
              "demo",
              "real"
            ],
            "title": "Environment",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/JobEventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "execution": {
            "additionalProperties": true,
            "title": "Execution",
            "type": "object"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "discovery",
              "refresh_number"
            ],
            "title": "Kind",
            "type": "string"
          },
          "next_attempt_at": {
            "format": "date-time",
            "title": "Next Attempt At",
            "type": "string"
          },
          "predecessor_job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predecessor Job Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "retry_cycle": {
            "minimum": 1.0,
            "title": "Retry Cycle",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "retry_wait",
              "completed",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "tribunal": {
            "title": "Tribunal",
            "type": "string"
          }
        },
        "required": [
          "id",
          "collection_id",
          "kind",
          "environment",
          "source",
          "tribunal",
          "status",
          "coverage",
          "reason",
          "cancel_requested",
          "attempt_count",
          "retry_cycle",
          "next_attempt_at",
          "created_at",
          "started_at",
          "finished_at",
          "criteria",
          "execution",
          "checkpoint",
          "cursor_invalid",
          "predecessor_job_id",
          "attempts",
          "events"
        ],
        "title": "JobDetailResponse",
        "type": "object"
      },
...
      "PaginationResponse_JobSummaryResponse_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginationResponse[JobSummaryResponse]",
        "type": "object"
      },
...
      "RefreshNumberCriteria": {
        "additionalProperties": false,
        "properties": {
          "hit_budget": {
            "default": 2000,
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Hit Budget",
            "type": "integer"
          },
          "page_size": {
            "default": 100,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "process_number": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Process Number",
            "type": "string"
          }
        },
        "required": [
          "process_number"
        ],
        "title": "RefreshNumberCriteria",
        "type": "object"
      },
      "RefreshNumberJobRequest": {
        "additionalProperties": false,
        "properties": {
          "criteria": {
            "$ref": "#/components/schemas/RefreshNumberCriteria"
          },
          "kind": {
            "const": "refresh_number",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "criteria"
        ],
        "title": "RefreshNumberJobRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API local de jobs, processos, representações, movimentos e presets.",
    "title": "AgroJud Radar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(queued|running|retry_wait|completed|partial|failed|cancelled)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(discovery|refresh_number)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse_JobSummaryResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Lista jobs persistidos",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "discovery": "#/components/schemas/DiscoveryJobRequest",
                    "refresh_number": "#/components/schemas/RefreshNumberJobRequest"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DiscoveryJobRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RefreshNumberJobRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Enfileira uma coleta tipada",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Consulta o estado e progresso de um job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Cancela ou solicita o cancelamento cooperativo",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/continue": {
      "post": {
        "operationId": "continue_job_api_v1_jobs__job_id__continue_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Continua uma coleta parcial que atingiu o orçamento",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/restart-scan": {
      "post": {
        "operationId": "restart_scan_api_v1_jobs__job_id__restart_scan_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Inicia nova varredura após invalidação do cursor",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/resume": {
      "post": {
        "operationId": "resume_job_api_v1_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Recurso não encontrado."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflito de estado ou capacidade indisponível."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Entrada inválida."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Banco de dados indisponível."
          }
        },
        "summary": "Retoma um job falho ou cancelado",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}